jason: PocketMode: Treat ProximitySensor values as int
The maximum range is reported as 5.000305 but the distance is just 5.0
which leads to isNear being always true. Cast them to int (the only
reported values are 0 and 5 anyway) to get the isNear value correct.

+++ pocketmode/src/org/lineageos/pocketmode/ProximitySensor.java
@@ -46,7 +46,7 @@ public class ProximitySensor implements SensorEventListener {
 
     @Override
     public void onSensorChanged(SensorEvent event) {
-        boolean isNear = event.values[0] < mSensor.getMaximumRange();
+        boolean isNear = (int) event.values[0] < (int) mSensor.getMaximumRange();
         if (FileUtils.isFileWritable(FPC_PROX_NODE)) {
             FileUtils.writeLine(FPC_PROX_NODE, isNear ? "1" : "0");
         }
